requirements.txt
groq
Flask
python-dotenv
requests
flask-socketio